Shipment 19: GCS

+++ exercises/models.py
@@ -8,6 +8,7 @@ from django.contrib.auth.models import User
 
 
 # Methods
+# if DEBUG True, will use localfile 
 
 def exercise_method_file_path(instance, filename):
     return os.path.join('methodes', instance.exercise_method.name, filename)
@@ -20,20 +21,23 @@ class ExerciseMethod(models.Model):
         # Delete all associated ExerciseMethodFile instances and their files
         for file_instance in self.files.all():
             if file_instance.file:
-                if os.path.isfile(file_instance.file.path):
-                    os.remove(file_instance.file.path)
+                if settings.DEBUG:
+                    # Local environment
+                    if os.path.isfile(file_instance.file.path):
+                        os.remove(file_instance.file.path)
+                else:
+                    # Production environment (Google Cloud Storage)
+                    file_instance.file.delete()
             file_instance.delete()
 
-        # Delete the method's directory
-        exercise_directory = os.path.join(
-            settings.MEDIA_ROOT, 'methodes', self.name)
-        if os.path.exists(exercise_directory):
-            shutil.rmtree(exercise_directory)
+        if settings.DEBUG:
+            # Delete the method's directory in local environment
+            exercise_directory = os.path.join(
+                settings.MEDIA_ROOT, 'methodes', self.name)
+            if os.path.exists(exercise_directory):
+                shutil.rmtree(exercise_directory)
 
         super().delete(*args, **kwargs)
-
-    def __str__(self):
-        return self.name
 
 
 class ExerciseMethodFile(models.Model):
@@ -43,6 +47,17 @@ class ExerciseMethodFile(models.Model):
     range = models.CharField(max_length=50)
     tempo = models.CharField(max_length=50)
     file = models.FileField(upload_to=exercise_method_file_path)
+
+    def delete(self, *args, **kwargs):
+        if self.file:
+            if settings.DEBUG:
+                # Local environment
+                if os.path.isfile(self.file.path):
+                    os.remove(self.file.path)
+            else:
+                # Production environment (Google Cloud Storage)
+                self.file.delete()
+        super().delete(*args, **kwargs)
 
     def __str__(self):
         return f"{self.exercise_method.name}_{self.range}_{self.direction}_{self.tempo}"
